+++ llm/cli.py
@@ -490,7 +490,7 @@ def cli():
 @click.option("-n", "--no-log", is_flag=True, help="Don't log to database")
 @click.option("--log", is_flag=True, help="Log prompt and response to the database")
 @click.option(
-    "-R", "--no-reasoning", is_flag=True, help="Don't display reasoning output"
+    "-R", "--hide-reasoning", is_flag=True, help="Hide reasoning output"
 )
 @click.option(
     "_continue",
@@ -542,7 +542,7 @@ def prompt(
     no_stream,
     no_log,
     log,
-    no_reasoning,
+    hide_reasoning,
     _continue,
     conversation_id,
     key,
@@ -911,7 +911,7 @@ def prompt(
                     )
                     await display_async_stream_events(
                         response.astream_events(),
-                        show_reasoning=not no_reasoning,
+                        show_reasoning=not hide_reasoning,
                     )
                     print("")
                 else:
@@ -946,7 +946,7 @@ def prompt(
             if should_stream:
                 display_stream_events(
                     response.stream_events(),
-                    show_reasoning=not no_reasoning,
+                    show_reasoning=not hide_reasoning,
                 )
                 print("")
             else:
@@ -1045,7 +1045,7 @@ def prompt(
 )
 @click.option("--no-stream", is_flag=True, help="Do not stream output")
 @click.option(
-    "-R", "--no-reasoning", is_flag=True, help="Don't display reasoning output"
+    "-R", "--hide-reasoning", is_flag=True, help="Hide reasoning output"
 )
 @click.option("--key", help="API key to use")
 @click.option(
@@ -1095,7 +1095,7 @@ def chat(
     param,
     options,
     no_stream,
-    no_reasoning,
+    hide_reasoning,
     key,
     database,
     tools,
@@ -1302,7 +1302,7 @@ def chat(
         argument_system_fragments = []
         display_stream_events(
             response.stream_events(),
-            show_reasoning=not no_reasoning,
+            show_reasoning=not hide_reasoning,
         )
         response.log_to_db(db)
         print("")